MAGENTO2-204: restructered service
